修改下载时增加下载状态记录的方式，修改下载数据返回单位

+++ src/main/java/com/devteam/mikufunbackend/entity/DownloadStatusV0.java
@@ -14,9 +14,9 @@ import lombok.experimental.Accessors;
 public class DownloadStatusV0 {
     String gid;
     String fileName;
-    int completedLength;
-    int fileSize;
-    int downloadSpeed;
-    int uploadSpeed;
+    double completedLength;
+    double fileSize;
+    double downloadSpeed;
+    double uploadSpeed;
     String status;
 }
